Restructure the rna result page to make it reusable.
