+++ ui/pages/Project/selectors.js
@@ -21,7 +21,6 @@ import {
   getMmeResultsByGuid, getMmeSubmissionsByGuid, getHasActiveSearchableSampleByFamily, getSelectableTagTypesByProject,
   getVariantTagsByGuid, getUserOptionsByUsername, getSamplesByFamily, getNotesByFamilyType,
   getSamplesGroupedByProjectGuid, getVariantTagNotesByFamilyVariants, getPhenotypeGeneScoresByIndividual,
-  getRnaSeqDataByIndividual,
 } from 'redux/selectors'
 
 import {
@@ -873,28 +872,5 @@ export const getIndividualPhenotypeGeneScores = createSelector(
         ]), []),
       ]), []),
     }), {})
-  ),
-)
-
-const getSpliceId = (row) => {
-  const { geneId, chrom, start, end, strand, type } = row
-  return `${geneId}-${chrom}-${start}-${end}-${strand}-${type}`
-}
-
-export const getRnaSeqSignificantJunctionData = createSelector(
-  getGenesById,
-  getRnaSeqDataByIndividual,
-  (genesById, rnaSeqDataByIndividual) => Object.entries(rnaSeqDataByIndividual).reduce(
-    (acc, [individualGuid, rnaSeqData]) => (rnaSeqData.spliceOutliers ? {
-      ...acc,
-      [individualGuid]: Object.values(rnaSeqData.spliceOutliers).flat().filter(({ isSignificant }) => isSignificant)
-        .sort((a, b) => a.pValue - b.pValue)
-        .map(row => ({
-          geneSymbol: (genesById[row.geneId] || {}).geneSymbol || row.geneId,
-          junctionLocus: `${row.chrom}:${row.start}-${row.end} ${row.strand}`,
-          idField: getSpliceId(row),
-          ...row,
-        })),
-    } : acc), {},
   ),
 )
